feat: adiciona atualização e remoção de usuários no CRUD

+++ src/main/java/com/priscila/usermanager/service/UserService.java
@@ -15,20 +15,38 @@ public class UserService {
         this.repository = repository;
     }
 
+    // CREATE
     public User create(User user) {
         return repository.save(user);
     }
 
+    // READ - LIST
     public List<User> findAll() {
         return repository.findAll();
     }
 
+    // READ - BY ID
     public User findById(Long id) {
-        return repository.findById(id).orElse(null);
+        return repository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
     }
 
-    public void delete(Long id) {
-        repository.deleteById(id);
+    // UPDATE
+    public User update(Long id, User user) {
+        User existing = repository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
 
+        existing.setNome(user.getNome());
+        existing.setEmail(user.getEmail());
+
+        return repository.save(existing);
+    }
+
+    //DELETE
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("Usuário não encontrado");
+        }
+        repository.deleteById(id);
     }
 }
